Revert release_version "fix" for mbed 2.
The online compiler requires mbed os 2 support. Now release_version has
been partially removed from the tools, we just need to return all targets
when any version is given. This doesn't really make sense, as now we just
claim to support all targets in all release_versions. However, it allows the
online compiler to find targets for mbed 2 projects, which we still need
to support online.

We need to come up with a mechanism for maintaining release_version
support in the online tools, this unfortunate patch is just to get
things working for the Mbed 6 release.